refactor: move ChangePreferenceView to preferences.views
